requirements.txt
awscli==1.25.82
black==22.8.0
botocore==1.27.81
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
decorator==5.1.1
distlib==0.3.4
docutils==0.16
filelock==3.7.1
html5lib==1.1
idna==3.3
jmespath==1.0.1
keyring==23.6.0
lockfile==0.12.2
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==20.9
pastel==0.2.1
pathspec==0.10.1
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry-core==1.0.8
ptyprocess==0.7.0
pyasn1==0.4.8
pylev==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.6.0
shellingham==1.4.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.1
urllib3==1.26.9
validators==0.18.2
virtualenv==20.15.1
weaviate-client==3.8.0
webencodings==0.5.1
